refactor: extract finite group implementation
